Created picking and displayed it onscreen for debug reasons.
Picking piped through CellRenderer to DrawPane to ApolloUI to
UIManagement. Information now needs to be processed by UIManagement
(processing the number of points) and then we'll send it over to Cell
for actual data edits. It seems like it's a lot of process for no
reason, but separating the data from the UI is absolutely vital.