Added a ton of classes that make up the framework for our event model.
Lots of questions about how we're doing this, so I'll leave an
explanation here.
Basically, Bootstrapper will implement two interfaces, one for UI and
one for data. The EventObject is stored wherever the events are held. It
is told to 'fire' an event and send a message to all classes that are
'listening'. Since it's an interface, each of them gets to pick their
response individually, thus creating a event model.